+++ src/main/ruby/truffleruby/core/process.rb
@@ -404,11 +404,12 @@ module Process
     # until respond_to? can be used to query the implementation of methods attached via FFI
     # atm respond_to returns true if a method is attached but not implemented on the platform
     uid =
-      if uid.kind_of?(String)
-        require 'etc'
-        Etc.getpwnam(uid).uid
+      if Truffle::Type.fits_into_long?(uid) ||
+        (tmp = Truffle::Type.rb_check_convert_type(uid, String, :to_str)).nil?
+        Truffle::Type.rb_num2uint(uid)
       else
-        Truffle::Type.coerce_to uid, Integer, :to_int
+        require 'etc'
+        Etc.getpwnam(tmp).uid
       end
 
     begin
@@ -436,11 +437,12 @@ module Process
 
   def self.egid=(gid)
     gid =
-      if gid.kind_of?(String)
-        require 'etc'
-        Etc.getgrnam(gid).gid
+      if Truffle::Type.fits_into_long?(gid) ||
+        (tmp = Truffle::Type.rb_check_convert_type(gid, String, :to_str)).nil?
+        Truffle::Type.rb_num2uint(gid)
       else
-        Truffle::Type.coerce_to gid, Integer, :to_int
+        require 'etc'
+        Etc.getgrnam(tmp).gid
       end
 
     Process::Sys.setegid gid